requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
bcrypt==4.2.0
beanie==1.26.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
confluent-kafka==2.8.0
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
debugpy==1.8.6
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
executing==2.1.0
fastapi==0.115.0
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.9.0
geographiclib==2.0
geopy==2.4.1
graphviz==0.20.3
h11==0.14.0
holidays==0.57
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kafka-python==2.0.3
kiwisolver==1.4.7
lazy-model==0.2.0
loguru==0.7.2
MarkupSafe==3.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
meteostat==1.6.8
mongoengine==0.29.1
motor==3.6.0
mpmath==1.3.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.4.1
nltk==3.9.1
numpy==2.1.1
odmantic==1.0.2
openai==1.54.1
opencv-python==4.11.0.86
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.9.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.38.6
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.48.2
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.70
ultralytics-thop==2.0.14
urllib3==2.2.3
uvicorn==0.31.0
wcwidth==0.2.13
win32-setctime==1.1.0
yarl==1.17.1
